Performance optimizations: `can` < `eval`
Instead of carefully checking if the object we work with has the shape
we want it to have, it's usually faster to just plunge ahead and deal
with the aftermath later. When you have your typical use case most of the time,
performance wins can be noticeable.

Note that `eval` statements here are constructed in such a way that if they fail,
they don't produce side effects.